+++ lib/testing-github-warden/executioner/executioner_helper.rb
@@ -52,9 +52,11 @@ module ExecutionerHelper
     end
   end
 
+  # Check if bug already was correctly commented
+  # @param commit_hash [String] commit hash string
+  # @param action_data [Hash] Github hook hash
+  # @return [Boolean]
   def bug_is_commented?(commit_hash, action_data)
-    return true if action_data.select { |action| action['bug_id'] }.empty?
-
     bug_id = action_data.find { |action| action['bug_id'] }['bug_id']
     comments = []
     5.times do |i|
@@ -75,5 +77,39 @@ module ExecutionerHelper
       comment['text'].include? commit_hash
     end
     !result.empty?
+  end
+
+  # Check if any action with bug should be done at all
+  # @param action_data [Array<Hash>] Github hook hash
+  # @param commit_hash [String] commit hash string
+  # @return [Boolean]
+  def bug_should_be_handled?(action_data, commit_hash)
+    return false unless hook_data_include_bug_number?(action_data)
+    return false unless bug_exists?(action_data)
+    return false if bug_is_commented?(commit_hash, action_data)
+
+    true
+  end
+
+  # Check if bug exists at all
+  # @param action_data [Array<Hash>] Github hook hash
+  # @return [True, False]
+  def bug_exists?(action_data)
+    bug_id = action_data.find { |action| action['bug_id'] }['bug_id']
+
+    exists = @bugzilla.bug_exists?(bug_id)
+    @logger.info("Bug for data: #{action_data} do not exists. Someone probably make a typo.") unless exists
+    exists
+  end
+
+  # Check if hook data contains bug id
+  # @param action_data [Array<Hash>] Github hook hash
+  # @return [Boolean]
+  def hook_data_include_bug_number?(action_data)
+    include = !action_data.select { |action| action['bug_id'] }.empty?
+
+    @logger.info("Github hook data: #{action_data} do not contains any bug_id. Skipping") unless include
+
+    include
   end
 end